refactor : 커뮤니티 코드 리팩토링
- 애플리케이션 서비스(유즈케이스)에서는 도메인 서비스들을 조합한 동작만 표현되도록 수정

+++ src/main/java/org/ezcode/codetest/domain/community/service/ReplyDomainService.java
@@ -5,6 +5,7 @@ import org.ezcode.codetest.domain.community.exception.CommunityExceptionCode;
 import org.ezcode.codetest.domain.community.model.Discussion;
 import org.ezcode.codetest.domain.community.model.Reply;
 import org.ezcode.codetest.domain.community.repository.ReplyRepository;
+import org.ezcode.codetest.domain.user.model.entity.User;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
@@ -17,7 +18,21 @@ public class ReplyDomainService {
 
 	private final ReplyRepository replyRepository;
 
-	public Reply createReply(Reply reply) {
+	public Reply createReply(Discussion discussion, User user, Long parentReplyId, String content) {
+
+		Reply parentReply = null;
+
+		if (parentReplyId != null) {
+			parentReply = getReplyById(parentReplyId);
+			validateDiscussionMatches(parentReply, discussion);
+		}
+
+		Reply reply = Reply.builder()
+			.discussion(discussion)
+			.user(user)
+			.parent(parentReply)
+			.content(content)
+			.build();
 
 		return replyRepository.save(reply);
 	}
@@ -33,17 +48,32 @@ public class ReplyDomainService {
 		return replyRepository.findAllRepliesByDiscussionId(discussion.getId(), pageable);
 	}
 
-	public Page<Reply> getRepliesByParentReplyId(Long parentReplyId, Pageable pageable) {
+	public Page<Reply> getRepliesByParentReplyId(Long parentReplyId, Discussion discussion, Pageable pageable) {
 
-		return replyRepository.findAllChildRepliesByParentReplyId(parentReplyId, pageable);
+		Reply parentReply = getReplyById(parentReplyId);
+		validateDiscussionMatches(parentReply, discussion);
+
+		return replyRepository.findAllChildRepliesByParentReplyId(parentReply.getId(), pageable);
 	}
 
-	public void modify(Reply reply, String content) {
+	public Reply modify(Long replyId, Discussion discussion, Long userId, String content) {
+
+		Reply reply = getReplyById(replyId);
+
+		validateDiscussionMatches(reply, discussion);
+		validateIsAuthor(reply, userId);
 
 		replyRepository.updateReply(reply, content);
+
+		return reply;
 	}
 
-	public void remove(Reply reply) {
+	public void remove(Long replyId, Discussion discussion, Long userId) {
+
+		Reply reply = getReplyById(replyId);
+
+		validateDiscussionMatches(reply, discussion);
+		validateIsAuthor(reply, userId);
 
 		replyRepository.deleteReply(reply);
 	}
